+++ src/ripe_rainbow/interactive/drivers.py
@@ -66,9 +66,13 @@ class InteractiveDriver(object):
         raise appier.NotImplementedError()
 
     def clear_storage(self):
+        # determines if the current page protocol is "compatible" with
+        # local storage (allow proper exception prevention)
+        storage_protocol = self.instance.execute_script("return !['data:'].includes(location.protocol);")
+
         self.clear_cookies()
-        self.clear_local_storage()
-        self.clear_session_storage()
+        if storage_protocol: self.clear_local_storage()
+        if storage_protocol: self.clear_session_storage()
 
     def clear_cookies(self):
         raise appier.NotImplementedError()